Add mutation for everything
Add  mutations to enable GTFS edit.
In order to mutate a GTFS detail, mutate a field with filters and add
the parameter you want to change with the new value.

Bonus:
Fixed issues on queries and added generic parameter handling for queries.
Also added feed_info for agency supporting it.

Change-Id: If8277a795e39c753c7c1ae8c5788e85ccc55f3e3